fix: Preserve native counter values in metrics callbacks
Separate recording limits from callback values, preserve unchanged unsigned Cocoa counters, and add native counter regression coverage. Include the Metrics PR link in the changelog.

+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/metrics/JvmSentryMetricsAdapter.kt
@@ -17,7 +17,7 @@ internal class JvmSentryMetricsAdapter(
         val params = SentryMetricsParameters.create(attributes.toJvmSentryAttributes())
         val metrics = provider()
         when (value) {
-            is SentryMetricValue.Counter -> metrics.count(name, value.value.toDouble(), unit, params)
+            is SentryMetricValue.Counter -> metrics.count(name, value.value, unit, params)
             is SentryMetricValue.Gauge -> metrics.gauge(name, value.value, unit, params)
             is SentryMetricValue.Distribution -> metrics.distribution(name, value.value, unit, params)
         }
